pyproject.toml
[project]
name = "workbench"
version = "0.1.0"
description = ""
authors = [
    { name = "Unusual Audio", email = "unusualaudio@protonmail.com" }
]
requires-python = ">=3.13,<4"
dependencies = [
    "jupyter (>=1.1.1,<2.0.0)",
    "pyusb (>=1.3.1,<2.0.0)",
    "pyvisa (>=1.15.0,<2.0.0)",
    "pyvisa-py (>=0.8.0,<0.9.0)",
    "psutil (>=7.0.0,<8.0.0)",
    "zeroconf (>=0.147.0,<0.148.0)",
    "hid (>=1.0.7,<2.0.0)",
    "dwfpy (>=1.2.0,<2.0.0)",
    "pyserial (>=3.5,<4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "scipy (>=1.15.3,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "pyvicp (>=1.1.0,<2.0.0)",
    "sympy (>=1.14.0,<2.0.0)",
    "si-prefix (>=1.3.3,<2.0.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "influxdb-client (>=1.49.0,<2.0.0)",
]

[project.scripts]
who-cares = "who_cares.__main__:main"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
